chore(client): fix 9 pre-existing clippy lints
Clean up warnings that were blocking `cargo clippy -- -D warnings`
on the client crate (unrelated to the answerback work).

Fixes:
- redundant_closure in app.rs
- collapsible_match on mouse handling in app.rs
- unnecessary_cast (2x) in caps/probe.rs
- new_without_default for InputQueue
- unnecessary_map_or in office/layout.rs
- unnecessary_map_or in office/state.rs
- new_without_default for ReconnectState
- needless_lifetimes in render/scene.rs

These were long-standing issues. Fixed as part of getting the
answerback branch fully green under strict linting.

+++ client/src/caps/probe.rs
@@ -66,7 +66,7 @@ impl Perform for ReplyParser {
                 let h = ps.next().unwrap_or(0);
                 let w = ps.next().unwrap_or(0);
                 if w > 0 && h > 0 {
-                    self.result.cell_px = Some((w as u16, h as u16));
+                    self.result.cell_px = Some((w, h));
                 }
             }
         }
